+++ CHANGELOG.md
@@ -33,6 +33,14 @@ There are now different themes that can be applied to Simple Calendar to change 
 
 I have tried to make sure that the light and dark themes are still easily usable for individuals with color blindness.
 
+### Client Settings
+
+Three new settings have been added to the Module Settings dialog of Foundry. These settings are [client settings](https://foundryvtt.com/article/settings/), which are saved in the browser and are applied on a per-browser basis.
+
+- **Open On Load**: This setting will open Simple Calendar's main interface when loading foundry. If the GM has disabled viewing the calendar for players, Simple Calendar will not show regardless of this setting. By default, this is set to true.
+- **Open In Compact View**: If checked anytime Simple Calendar is opened (On Load, Clicking the button or through a macro) it will be opened in the compact view. By default, this is set to false.
+- **Remember Position**: If checked Simple Calendar will remember where it was placed on the screen and position itself there everytime it is opened. By default, this is set to true.
+
 ### Improved Date/Time Selector
 
 The date/time selector that is used when choosing a date for notes or specifying the month of a Season or the sunrise/sunset time of a season has been improved.